Fix building twice in a row
Some of the changes performed during both the autoreconf and
clean steps, while not compromising the idempotency of the
build process in its entirety, should either be ignored or
can't even be represented as regular diffs at all.

We don't want stray changes to end up in the automatic Debian
patch anyway, so just ignore all of them.