Introduce safeguards against unnoticed SONAME bumps
Instead of carelessly picking up usr/lib/*/libcattle-1.0.so.*
for installation, be a bit more considerate and include the
current SONAME in the pattern.

Doing so means that, should upstream bump the SONAME due to
incompatible changes without the Debian maintainer noticing,
the package would immediately FTBFS instead of introducing
undetected breakages.

Suggested-by: Pino Toscano <pino@debian.org>