install: Expand almost all wildcards
This ensures that new files introduced upstream will not slip
into a binary package without dh_missing notifying us.

The ${DEB_HOST_MULTIARCH} variable, which is supported by
debhelper starting with compat level 13, is used liberally
to handle paths that change based on the architecture.

There is only one case in which wildcards are still used: the
version part of the filename for the library itself changes
every release, and it would be annoying to have to keep up.

These wildcards are already written in a way that will catch
unexpected incompatible soname bumps, so in practice that's
not going to be a problem.